feat: redesign recovery flow
