+++ src/components/common/SocialSigner/PasswordRecovery.tsx
@@ -1,15 +1,5 @@
 import { MPC_WALLET_EVENTS } from '@/services/analytics/events/mpcWallet'
-import {
-  Typography,
-  FormControlLabel,
-  Checkbox,
-  Button,
-  Box,
-  Divider,
-  Grid,
-  LinearProgress,
-  FormControl,
-} from '@mui/material'
+import { Typography, FormControlLabel, Checkbox, Button, Box, Divider, FormControl, Avatar } from '@mui/material'
 import { useState } from 'react'
 import Track from '@/components/common/Track'
 import { FormProvider, useForm } from 'react-hook-form'
@@ -60,60 +50,51 @@ export const PasswordRecovery = ({
   return (
     <FormProvider {...formMethods}>
       <form onSubmit={handleSubmit(onSubmit)}>
-        <Grid container justifyContent="center" alignItems="center">
-          <Grid item xs={12} md={5} p={2}>
-            <Typography variant="h2" fontWeight="bold" mb={3}>
-              Verify your account
-            </Typography>
-            <Box bgcolor="background.paper" borderRadius={1}>
-              <LinearProgress
-                color="secondary"
-                sx={{ borderTopLeftRadius: '6px', borderTopRightRadius: '6px', opacity: isDisabled ? 1 : 0 }}
-              />
-              <Box p={4}>
-                <Typography variant="h6" fontWeight="bold" mb={0.5}>
-                  Enter security password
-                </Typography>
-                <Typography>
-                  This browser is not registered with your Account yet. Please enter your recovery password to restore
-                  access to this Account.
-                </Typography>
-              </Box>
-              <Divider />
-              <Box className={css.passwordWrapper}>
-                <FormControl fullWidth>
-                  <PasswordInput
-                    name="password"
-                    label="Recovery password"
-                    helperText={formState.errors['password']?.message}
-                    disabled={isDisabled}
-                    required
-                  />
-                </FormControl>
-                <FormControlLabel
-                  disabled={isDisabled}
-                  control={
-                    <Checkbox checked={storeDeviceFactor} onClick={() => setStoreDeviceFactor((prev) => !prev)} />
-                  }
-                  label="Do not ask again on this device"
-                />
-                {error && <ErrorMessage className={css.loginError}>{error}</ErrorMessage>}
-              </Box>
-
-              <Divider />
-              <Box p={4} display="flex" justifyContent="space-between">
-                <Button variant="outlined" onClick={onBack}>
-                  Back
-                </Button>
-                <Track {...MPC_WALLET_EVENTS.RECOVER_PASSWORD}>
-                  <Button variant="contained" type="submit" disabled={isDisabled}>
-                    Submit
-                  </Button>
-                </Track>
-              </Box>
+        <Box>
+          <Box p={4}>
+            <Box display="flex" flexDirection="row" gap={1} alignItems="center" mb={0.5}>
+              <Avatar className={css.dot}>
+                <Typography variant="body2">{2}</Typography>
+              </Avatar>
+              <Typography variant="h6" fontWeight="bold">
+                Enter recovery password
+              </Typography>
             </Box>
-          </Grid>
-        </Grid>
+            <Typography variant="body2" pl={'28px'}>
+              Please enter your recovery password.
+            </Typography>
+          </Box>
+          <Divider />
+          <Box className={css.passwordWrapper}>
+            <FormControl fullWidth>
+              <PasswordInput
+                name="password"
+                label="Recovery password"
+                helperText={formState.errors['password']?.message}
+                disabled={isDisabled}
+                required
+              />
+            </FormControl>
+            <FormControlLabel
+              disabled={isDisabled}
+              control={<Checkbox checked={storeDeviceFactor} onClick={() => setStoreDeviceFactor((prev) => !prev)} />}
+              label="Do not ask again on this device"
+            />
+            {error && <ErrorMessage className={css.loginError}>{error}</ErrorMessage>}
+          </Box>
+
+          <Divider />
+          <Box p={4} display="flex" justifyContent="space-between">
+            <Button variant="outlined" onClick={onBack}>
+              Back
+            </Button>
+            <Track {...MPC_WALLET_EVENTS.RECOVER_PASSWORD}>
+              <Button variant="contained" type="submit" disabled={isDisabled}>
+                Submit
+              </Button>
+            </Track>
+          </Box>
+        </Box>
       </form>
     </FormProvider>
   )
